🚚 Rename endpoints

+++ src/main/java/com/dodatabase/demo/controller/WishController.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,7 +25,7 @@ public class WishController {
   private final WishService wishService;
 
   @GetMapping("")
-  public String list(Model model) {
+  public String wishList(Model model) {
     List<WishResponse> wishResponses = wishService.findWishes();
     model.addAttribute("wishes", wishResponses);
     return "html/wish/list";
@@ -32,7 +33,7 @@ public class WishController {
 
   @PostMapping("")
   @ResponseBody
-  public ResponseEntity<Void> createMovie(@RequestBody WishRequest wishRequest) {
+  public ResponseEntity<Void> createWish(@RequestBody WishRequest wishRequest) {
     if (wishService.findById(wishRequest.getId()).isPresent()) {
       return ResponseEntity.status(HttpStatus.CONFLICT).build();
     }
@@ -41,10 +42,18 @@ public class WishController {
     return ResponseEntity.status(HttpStatus.CREATED).build();
   }
 
-  @DeleteMapping("")
+  @DeleteMapping("/{id}")
   @ResponseBody
-  public ResponseEntity<Void> removeMovie(@RequestBody String id) {
-    wishService.deleteById(id);
-    return ResponseEntity.status(HttpStatus.OK).build();
+  public ResponseEntity<Void> removeWish(@PathVariable String id) {
+    try {
+      if (!wishService.findById(id).isPresent()) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build(); // ID가 존재하지 않는 경우
+      }
+      wishService.deleteById(id);
+      return ResponseEntity.status(HttpStatus.OK).build();
+    } catch (Exception e) {
+      return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build(); // 서버 내부 오류 처리
+    }
   }
+
 }
